src/requirements.txt
django~=3.0.5
djangorestframework~=3.11.0
django-rest-swagger~=2.2.0
drf-yasg~=1.17.0
mozilla-django-oidc~=1.2.3
whitenoise~=5.0.1
dj-database-url~=0.5.0
requests~=2.23.0
django-watchman~=1.1.0
django-webpack-loader~=0.6.0
django-safedelete~=0.5.4
jsonfield~=3.1.0
django-redirect-to-non-www~=0.1.1
drf-api-tracking~=1.6.0
django-filter~=2.2.0
channels~=2.4.0
channels-redis~=2.4.2
twilio~=6.45.4
